Add useCallback
