+++ src/components/cart-dropdown/cart-dropdown.component.tsx
@@ -1,0 +1,36 @@
+import { useSelector } from "react-redux";
+import { useNavigate } from "react-router-dom";
+
+import Button from "../button/button.component";
+import CartItem from "../cart-item/cart-item.component";
+import { selectCartItems } from "../../store/cart/cart.selector";
+import {
+  CartDropdownContainer,
+  EmptyMessage,
+  CartItems,
+} from "./cart-dropdown.styles";
+import { useCallback, useMemo } from "react";
+
+const CartDropdown = () => {
+  const cartItems = useSelector(selectCartItems);
+  const navigate = useNavigate();
+
+  const goToCheckoutHandler = useCallback(() => {
+    navigate("/checkout");
+  }, []);
+
+  return (
+    <CartDropdownContainer>
+      <CartItems>
+        {cartItems.length ? (
+          cartItems.map((item) => <CartItem key={item.id} cartItem={item} />)
+        ) : (
+          <EmptyMessage>Your cart is empty</EmptyMessage>
+        )}
+      </CartItems>
+      <Button onClick={goToCheckoutHandler}>GO TO CHECKOUT</Button>
+    </CartDropdownContainer>
+  );
+};
+
+export default CartDropdown;
